template: add option to disable sql escaping on a template property
The patch adds a field option bDisableSQL to the structure templateEntry.
It is useful when we wan't to be responsible for SQL escaping on
particular property, for instance to be able to insert NULL values.
From a user point of view, when option nosqlescape="on" is set on a
property, the SQL escaping that was set on the template is ignored.

(cherry picked from commit f42f8049056a9c20f02c77464cdc64cc904faa9f)