imuxsock: make special parser strip tag/msg space separator
The msg part of a message parsed using the imuxsock special parser
always starts with a space character. The patch adds new module option
specialparser.removemsgfirstspace allowing to strip that space. This
behavior may also be controlled on the input declaration through option
specialparser.removemsgfirstspace.

This should be configurable and disabled by default in order to preserve
compatibility. This behavior makes sence for consistency when the new
pmrfc3164 option remove.MsgFirstSpace is enabled.